Rewrite the main network folder tests (basically check they call the helper functions buildURL and makeRequest with the correct params). Move helper functions to separate folder. AMAAS-950
