+++ src/utils/network/index.js
@@ -1,14 +1,5 @@
-require('dotenv').config()
-import request from 'superagent'
-import { endpoint, userPoolId, clientAppId } from '../../config.js'
-import fs from 'fs'
-import { AuthenticationDetails, CognitoUser, CognitoUserPool } from 'amazon-cognito-identity-js'
-import expandTilde from 'expand-tilde'
+import * as utils from './utils'
 
-const userPool = new CognitoUserPool({
-  UserPoolId: userPoolId,
-  ClientId: clientAppId
-})
 let stage = 'staging'
 let apiVersion = 'v1.0'
 let token
@@ -34,199 +25,6 @@ export function configureAuth(config) {
   return
 }
 
-export function getEndpoint() {
-  switch (stage) {
-    case 'staging':
-      return `${endpoint.staging}/staging`
-    case 'prod':
-      return `${endpoint.prod}/${apiVersion}`
-    default:
-      console.warn(`Unknown stage variable: ${stage}. Defaulting to /prod`)
-      return `${endpoint.prod}/${apiVersion}`
-  }
-}
-
-function isNode() {
-  try {
-    return Object.prototype.toString.call(global.process) === '[object process]'
-  } catch(e) {
-    return false
-  }
-}
-
-export function authenticate() {
-  let injectedResolve
-  let injectedReject
-  return new Promise((resolve, reject) => {
-    injectedResolve = resolve
-    injectedReject = reject
-    let path
-    if (credPath) {
-      path = credPath
-    } else {
-      path = `${expandTilde('~')}/amaas.js`
-    }
-    console.log(`Reading credentials from ${path}`)
-    fs.readFile(path, (error, data)=> {
-      if (error) {
-        return injectedReject(error)
-    }
-      const Username = JSON.parse(data).username
-      const Password = JSON.parse(data).password
-      const authenticationDetails = new AuthenticationDetails({
-        Username,
-        Password
-      })
-      const cognitoUser = new CognitoUser({
-        Username,
-        Pool: userPool
-      })
-      console.log('Starting authentication...')
-      cognitoUser.authenticateUser(authenticationDetails, {
-        onSuccess: res => injectedResolve(res.getIdToken().getJwtToken()),
-        onFailure: err => injectedReject(err)
-      })
-    })
-  })
-}
-
-export function getToken() {
-  if (token && token.length > 0) {
-    return Promise.resolve(token)
-  }
-  let injectedResolve
-  let injectedReject
-  return new Promise((resolve, reject) => {
-    injectedResolve = resolve
-    injectedReject = reject
-    switch (stage) {
-      case 'staging':
-      case 'prod':
-        const cognitoUser = userPool.getCurrentUser()
-        if (!cognitoUser) {
-          if (isNode()) {
-            console.warn('No user in storage, attempting to authenticate...')
-            authenticate()
-              .then(res => injectedResolve(res))
-              .catch(err => injectedReject(err))
-          } else {
-            injectedReject('Unauthorized, please re-authenticate')
-          }
-        } else {
-          cognitoUser.getSession((err, session) => {
-            if (session) {
-              console.log('getSession success')
-              injectedResolve(session.getIdToken().getJwtToken())
-            } else {
-              if (isNode()) {
-                console.warn('getSession failure, attempting to authenticate')
-                  authenticate()
-                  .then(res => injectedResolve(res))
-                  .catch(err => injectedReject(err))
-              } else {
-                injectedReject('Unauthorized, please re-authenticate')
-              }
-            }
-          })
-        }
-        break
-      default:
-        injectedReject('Missing Authorization')
-    }
-  })
-}
-
-/***
- * !This is an internal function that should not be called by the end user!
-
- * Builds a URL for HTTP request
- * @param {object} anonymous: anonyous object with arguments:
- * @param {string} AMaaSClass: class being requested/sent (e.g. Transaction) (required)
- * @param {string} AMId: Asset Manager Id (required)
- * @param {string} resourceId: Id of the resource being requested (e.g. book_id)
-*/
-export function buildURL({ AMaaSClass, AMId, resourceId }) {
-  let baseURL = ''
-  switch (AMaaSClass) {
-    case 'book':
-      baseURL = `${getEndpoint()}/book/books`
-      break
-    case 'parties':
-      baseURL = `${getEndpoint()}/party/parties`
-      break
-    case 'assetManagers':
-      baseURL = `${getEndpoint()}/assetmanager/asset-managers`
-      break
-    case 'assets':
-      baseURL = `${getEndpoint()}/asset/assets`
-      break
-    case 'positions':
-      baseURL = `${getEndpoint()}/transaction/positions`
-      break
-    case 'allocations':
-      baseURL = `${getEndpoint()}/transaction/allocations`
-      break
-    case 'monitorItems':
-      baseURL = `${getEndpoint()}/monitor/items`
-      break
-    case 'monitorEvents':
-      baseURL = `${getEndpoint()}/monitor/events`
-      break
-    case 'netting':
-      baseURL = `${getEndpoint()}/transaction/netting`
-      break
-    case 'relationships':
-      baseURL = `${getEndpoint()}/assetmanager/asset-manager-relationships`
-      break
-    case 'transactions':
-      baseURL = `${getEndpoint()}/transaction/transactions`
-      break
-    case 'corporateActions':
-      baseURL = `${getEndpoint()}/corporateaction/corporate-actions`
-      break
-    default:
-      throw new Error(`Invalid class type: ${AMaaSClass}`)
-  }
-  if (!AMId) {
-    return `${baseURL}`
-  } else if (!resourceId) {
-    return `${baseURL}/${AMId}`
-  } else {
-    return `${baseURL}/${AMId}/${resourceId}`
-  }
-}
-
-export function setAuthorization() {
-  switch (stage) {
-    case 'staging':
-    case 'prod':
-    default:
-      return 'Authorization'
-  }
-}
-
-export function makeRequest({ method, url, data, query }) {
-  return getToken()
-    .then(res => {
-      switch (method) {
-        case 'GET':
-          return request.get(url).set(setAuthorization(), res).query(query)
-        case 'SEARCH':
-          return request.get(url).set(setAuthorization(), res).query(data)
-        case 'POST':
-          return request.post(url).send(data).set(setAuthorization(), res).query(query)
-        case 'PUT':
-          return request.put(url).send(data).set(setAuthorization(), res).query({ camelcase: true })
-        case 'PATCH':
-          return request.patch(url).send(data).set(setAuthorization(), res).query({ camelcase: true })
-        case 'DELETE':
-          return request.delete(url).set(setAuthorization(), res).query({ camelcase: true })
-        default:
-      }
-    })
-    .catch(err => Promise.reject(err))
-}
-
 /***
  * !This is an internal function that should not be called by the end user!
  * !Wrapper functions are exposed for the individual asset classes for consumption!
@@ -239,12 +37,11 @@ export function makeRequest({ method, url, data, query }) {
  * @param {object} [query] - Additional query parameters
 */
 export function retrieveData({ AMaaSClass, AMId, resourceId, query }, callback) {
-  console.log(buildURL)
   let url
   let data = {}
   // If resourceId is supplied, append to url. Otherwise, return all data for AMId
   try {
-    url = buildURL({ AMaaSClass, AMId, resourceId })
+    url = utils.buildURL({ AMaaSClass, AMId, resourceId })
   } catch (e) {
     if (typeof callback !== 'function') {
       return Promise.reject(e)
@@ -259,7 +56,7 @@ export function retrieveData({ AMaaSClass, AMId, resourceId, query }, callback) 
       }
     }
   }
-  let promise = makeRequest({ method: 'GET', url, query: { ...data, camelcase: true } })
+  let promise = utils.makeRequest({ method: 'GET', url, query: { ...data, camelcase: true } })
   // let promise = request.get(url).set('x-api-key', token).query({ camelcase: true })
   if (typeof callback !== 'function') {
     // return promise if callback is not provided
@@ -295,7 +92,7 @@ export function retrieveData({ AMaaSClass, AMId, resourceId, query }, callback) 
 export function insertData({ AMaaSClass, AMId, resourceId, data, queryParams }, callback) {
   let url
   try {
-    url = buildURL({
+    url = utils.buildURL({
       AMaaSClass,
       AMId,
       resourceId
@@ -319,7 +116,7 @@ export function insertData({ AMaaSClass, AMId, resourceId, data, queryParams }, 
     }
     Object.assign(query, queryParams)
   }
-  let promise = makeRequest({ method: 'POST', url, data, query })
+  let promise = utils.makeRequest({ method: 'POST', url, data, query })
   // let promise = request.post(url).send(data).set('x-api-key', token).query({ camelcase: true })
   if (typeof callback !== 'function') {
     // return promise if callback is not provided
@@ -339,7 +136,7 @@ export function insertData({ AMaaSClass, AMId, resourceId, data, queryParams }, 
 export function putData({ AMaaSClass, AMId, resourceId, data }, callback) {
   let url
   try {
-    url = buildURL({
+    url = utils.buildURL({
       AMaaSClass,
       AMId,
       resourceId
@@ -355,7 +152,7 @@ export function putData({ AMaaSClass, AMId, resourceId, data }, callback) {
     url,
     json: data
   }
-  let promise = makeRequest({ method: 'PUT', url, data })
+  let promise = utils.makeRequest({ method: 'PUT', url, data })
   // let promise = request.put(url).send(data).set('x-api-key', token).query({ camelcase: true })
   if (typeof callback !== 'function') {
     // return promise if callback is not provided
@@ -375,7 +172,7 @@ export function putData({ AMaaSClass, AMId, resourceId, data }, callback) {
 export function patchData({ AMaaSClass, AMId, resourceId, data }, callback) {
   let url
   try {
-    url = buildURL({
+    url = utils.buildURL({
       AMaaSClass,
       AMId,
       resourceId
@@ -391,13 +188,13 @@ export function patchData({ AMaaSClass, AMId, resourceId, data }, callback) {
     url,
     json: data
   }
-  let promise = makeRequest({ method: 'PATCH', url, data })
+  let promise = utils.makeRequest({ method: 'PATCH', url, data })
   // let promise = request.patch(url).send(data).set('x-api-key', token).query({ camelcase: true })
   if (typeof callback !== 'function') {
     // return promise if callback is not provided
     return promise.then(response => response.body)
   }
-  promise.end((response, error) => {
+  promise.then((response, error) => {
     let body
     if (response) body = response.body
     _networkCallback(error, response, body, callback)
@@ -411,7 +208,7 @@ export function patchData({ AMaaSClass, AMId, resourceId, data }, callback) {
 export function deleteData({ AMaaSClass, AMId, resourceId }, callback) {
   let url
   try {
-    url = buildURL({
+    url = utils.buildURL({
       AMaaSClass,
       AMId,
       resourceId
@@ -423,7 +220,7 @@ export function deleteData({ AMaaSClass, AMId, resourceId }, callback) {
     callback(e)
     return
   }
-  let promise = makeRequest({ method: 'DELETE', url })
+  let promise = utils.makeRequest({ method: 'DELETE', url })
   // let promise = request.delete(url).set('x-api-key', token).query({ camelcase: true })
   if (typeof callback !== 'function') {
     // return promise if callback is not provided
@@ -449,7 +246,7 @@ export function deleteData({ AMaaSClass, AMId, resourceId }, callback) {
 export function searchData({ AMaaSClass, AMId, query }, callback) {
   let url
   try {
-    url = buildURL({
+    url = utils.buildURL({
       AMaaSClass,
       AMId
     })
@@ -466,7 +263,7 @@ export function searchData({ AMaaSClass, AMId, query }, callback) {
       data[q] = query[q].join()
     }
   }
-  let promise = makeRequest({ method: 'SEARCH', url, data })
+  let promise = utils.makeRequest({ method: 'SEARCH', url, data })
   if (typeof callback !== 'function') {
     // return promise if callback is not provided
     return promise.then(response => response.body)
