+++ src/utils/network/network.test.js
@@ -1,132 +1,141 @@
-import nock from 'nock'
-import {
-  buildURL,
-  retrieveData,
-  insertData,
-  searchData,
-  putData,
-  patchData,
-  deleteData,
-  getEndpoint
-} from './'
-import * as funcs from './'
+import * as network from './'
+import * as utils from './utils'
 import * as api from '../../exports/api'
+
+utils.buildURL = jest.fn()
+utils.makeRequest = jest.fn()
 
 api.config({
   stage: 'staging'
 })
 
-describe.only('buildURL function', () => {
-  it('should throw if no class supplied', () => {
-    function tester() {
-      buildURL({})
-    }
-    expect(tester).toThrowError('Invalid class type: undefined')
-  })
-  it('should not build past class if resourceId is supplied but not AMId', () => {
-    const testParams = {
-      AMaaSClass: 'book',
-      AMId: undefined,
-      resourceId: 'testResource'
-    }
-    const expectedURL = `${getEndpoint()}/book/books`
-    expect(buildURL(testParams)).toEqual(expectedURL)
-  })
-  it('should build correctly if all parameters are specified', () => {
-    const testParams = {
-      AMaaSClass: 'book',
-      AMId: 'testAMId',
-      resourceId: 'testResource'
-    }
-    const expectedURL = `${getEndpoint()}/book/books/testAMId/testResource`
-    expect(buildURL(testParams)).toEqual(expectedURL)
-  })
-})
-
 describe('retrieveData', () => {
-  afterAll(() => {
-    nock.cleanAll()
+  beforeAll(() => {
+    utils.buildURL.mockImplementation(() => 'testURL')
+    utils.makeRequest.mockImplementation(() => Promise.resolve({ body: 'testBody' }))
   })
   const testParams = {
     AMaaSClass: 'book',
-    AMId: '1234'
+    AMId: 1234
   }
-  it('should hit the correct endpoint', callback => {
-    const scope = nock(getEndpoint())
-      .get('/book/books/1234?camelcase=true')
-      .reply(200, {
-        param1: 'testBody'
-      })
-    retrieveData(testParams, (error, result) => {
-      expect(result).toEqual({ param1: 'testBody' })
+  it('calls buildURL with correct params', callback => {
+    network.retrieveData(testParams, (error, result) => {
+      expect(utils.buildURL).toHaveBeenCalledWith({ AMaaSClass: 'book', AMId: 1234 })
       callback()
     })
   })
-  it('should receive the correct HTTP status code', callback => {
-    const scope = nock(getEndpoint())
-      .get('/book/books/1234?camelcase=true')
-      .reply(501)
-    retrieveData(testParams, (error, result) => {
-      expect(error.status).toEqual(501)
+  it('calls makeRequest with correct params', callback => {
+    network.retrieveData(testParams, (error, result) => {
+      expect(utils.makeRequest).toHaveBeenCalledWith({ method: 'GET', url: 'testURL', query: { camelcase: true } })
       callback()
     })
   })
   it('should return a promise if callback is not provided', () => {
-    let promise = retrieveData(testParams).catch(error => {})
+    let promise = network.retrieveData(testParams)
     expect(promise).toBeInstanceOf(Promise)
   })
 })
 
 describe('insertData', () => {
-  afterAll(() => {
-    nock.cleanAll()
+  beforeAll(() => {
+    utils.buildURL.mockImplementation(() => 'testURL')
+    utils.makeRequest.mockImplementation(() => Promise.resolve({ body: 'testBody' }))
   })
   const testParams = {
     AMaaSClass: 'book',
-    AMId: '1234',
+    AMId: 1234,
     data: {
       price: 20
     }
   }
-  it('should build the correct url and POST to it', callback => {
-    const scope = nock(getEndpoint())
-      .post('/book/books/1234?camelcase=true')
-      .reply(200, {
-        param1: 'testResponse'
-      })
-    insertData(testParams, (error, result) => {
-      expect(result).toEqual({ param1: 'testResponse' })
+  it('should call buildURL with correct params', callback => {
+    network.insertData(testParams, (error, result) => {
+      expect(utils.buildURL).toHaveBeenCalledWith({ AMaaSClass: 'book', AMId: 1234 })
+      callback()
+    })
+  })
+  it('should call makeRequest with correct params', callback => {
+    network.insertData(testParams, (error, result) => {
+      expect(utils.makeRequest).toHaveBeenCalledWith({ method: 'POST', url: 'testURL', data: { price: 20 }, query: { camelcase: true } })
       callback()
     })
   })
   it('should return a promise if callback is not provided', () => {
-    let promise = insertData(testParams).catch(error => {})
+    let promise = network.insertData(testParams)
     expect(promise).toBeInstanceOf(Promise)
   })
 })
 
 describe('searchData', () => {
+  beforeAll(() => {
+    utils.buildURL.mockImplementation(() => 'testURL')
+    utils.makeRequest.mockImplementation(() => Promise.resolve({ body: 'testBody' }))
+  })
   const query = {
     assetManagerIds: [1, 2]
   }
+  it('should call buildURL with correct params', done => {
+    network.searchData({ AMaaSClass: 'monitorItems', AMId: 1, query }, (error, result) => {
+      expect(utils.buildURL).toHaveBeenCalledWith({ AMaaSClass: 'monitorItems', AMId: 1 })
+      done()
+    })
+  })
+  it('should call makeRequest with correct params', done => {
+    network.searchData({ AMaaSClass: 'monitorItems', AMId: 1, query }, (error, result) => {
+      expect(utils.makeRequest).toHaveBeenCalledWith({ method: 'SEARCH', url: 'testURL', data: { camelcase: true, assetManagerIds: '1,2' } })
+      done()
+    })
+  })
   it('should return a promise if callback is not provided', () => {
-    let promise = searchData({}).then(res => {}).catch(error => {})
+    let promise = network.searchData({})
     expect(promise).toBeInstanceOf(Promise)
   })
 })
 
 describe('putData', () => {
-  let params = { AMaaSClass: 'positions'}
+  beforeAll(() => {
+    utils.buildURL.mockImplementation(() => 'testURL')
+    utils.makeRequest.mockImplementation(() => Promise.resolve({ body: 'testBody' }))
+  })
+  it('should call buildURL with correct params', done => {
+    network.putData({ AMaaSClass: 'parties', AMId: 1, resourceId: 'testID', data: { test: 'testData' } }, (error, result) => {
+      expect(utils.buildURL).toHaveBeenCalledWith({ AMaaSClass: 'parties', AMId: 1, resourceId: 'testID' })
+      done()
+    })
+  })
+  it('should call makeRequest with correct params', done => {
+    network.putData({ AMaaSClass: 'parties', AMId: 1, resourceId: 'testID', data: { test: 'testData' } }, (error, result) => {
+      expect(utils.makeRequest).toHaveBeenCalledWith({ method: 'PUT', url: 'testURL', data: { test: 'testData' } })
+      done()
+    })
+  })
   it('should return a promise if callback is not provided', () => {
-    let promise = putData(params).catch(error => {})
+    let promise = network.putData({ AMaaSClass: 'parties', AMId: 1, resourceId: 'testID', data: { test: 'testData' } })
     expect(promise).toBeInstanceOf(Promise)
   })
 })
 
 describe('patchData', () => {
-  let params = { AMaaSClass: 'positions'}
+  let params
+  beforeAll(() => {
+    params = { AMaaSClass: 'positions', AMId: 1, resourceId: 'testID', data: { change: 'changed' } }
+    utils.buildURL.mockImplementation(() => 'testURL')
+    utils.makeRequest.mockImplementation(() => Promise.resolve({ body: 'testBody' }))
+  })
+  it('should call buildURL with the correct params', done => {
+    network.patchData(params, (error, result) => {
+      expect(utils.buildURL).toHaveBeenCalledWith({ AMaaSClass: 'positions', AMId: 1, resourceId: 'testID' })
+      done()
+    })
+  })
+  it('should call makeRequest with the correct params', done => {
+    network.patchData(params, (error, result) => {
+      expect(utils.makeRequest).toHaveBeenCalledWith({ method: 'PATCH', url: 'testURL', data: { change: 'changed' } })
+      done()
+    })
+  })
   it('should return a promise if callback is not provided', () => {
-    let promise = patchData(params).catch(error => {})
+    let promise = network.patchData(params)
     expect(promise).toBeInstanceOf(Promise)
   })
 })
