+++ src/utils/network/utils.js
@@ -1,0 +1,203 @@
+import request from 'superagent'
+import { endpoint, userPoolId, clientAppId } from '../../config.js'
+import fs from 'fs'
+import { AuthenticationDetails, CognitoUser, CognitoUserPool } from 'amazon-cognito-identity-js'
+import expandTilde from 'expand-tilde'
+
+const userPool = new CognitoUserPool({
+  UserPoolId: userPoolId,
+  ClientId: clientAppId
+})
+
+export function getEndpoint() {
+  switch (stage) {
+    case 'staging':
+      return `${endpoint.staging}/staging`
+    case 'prod':
+      return `${endpoint.prod}/${apiVersion}`
+    default:
+      console.warn(`Unknown stage variable: ${stage}. Defaulting to /prod`)
+      return `${endpoint.prod}/${apiVersion}`
+  }
+}
+
+function isNode() {
+  try {
+    return Object.prototype.toString.call(global.process) === '[object process]'
+  } catch(e) {
+    return false
+  }
+}
+
+export function authenticate() {
+  let injectedResolve
+  let injectedReject
+  return new Promise((resolve, reject) => {
+    injectedResolve = resolve
+    injectedReject = reject
+    let path
+    if (credPath) {
+      path = credPath
+    } else {
+      path = `${expandTilde('~')}/amaas.js`
+    }
+    console.log(`Reading credentials from ${path}`)
+    fs.readFile(path, (error, data)=> {
+      if (error) {
+        return injectedReject(error)
+    }
+      const Username = JSON.parse(data).username
+      const Password = JSON.parse(data).password
+      const authenticationDetails = new AuthenticationDetails({
+        Username,
+        Password
+      })
+      const cognitoUser = new CognitoUser({
+        Username,
+        Pool: userPool
+      })
+      console.log('Starting authentication...')
+      cognitoUser.authenticateUser(authenticationDetails, {
+        onSuccess: res => injectedResolve(res.getIdToken().getJwtToken()),
+        onFailure: err => injectedReject(err)
+      })
+    })
+  })
+}
+
+export function getToken() {
+  if (token && token.length > 0) {
+    return Promise.resolve(token)
+  }
+  let injectedResolve
+  let injectedReject
+  return new Promise((resolve, reject) => {
+    injectedResolve = resolve
+    injectedReject = reject
+    switch (stage) {
+      case 'staging':
+      case 'prod':
+        const cognitoUser = userPool.getCurrentUser()
+        if (!cognitoUser) {
+          if (isNode()) {
+            console.warn('No user in storage, attempting to authenticate...')
+            authenticate()
+              .then(res => injectedResolve(res))
+              .catch(err => injectedReject(err))
+          } else {
+            injectedReject('Unauthorized, please re-authenticate')
+          }
+        } else {
+          cognitoUser.getSession((err, session) => {
+            if (session) {
+              console.log('getSession success')
+              injectedResolve(session.getIdToken().getJwtToken())
+            } else {
+              if (isNode()) {
+                console.warn('getSession failure, attempting to authenticate')
+                  authenticate()
+                  .then(res => injectedResolve(res))
+                  .catch(err => injectedReject(err))
+              } else {
+                injectedReject('Unauthorized, please re-authenticate')
+              }
+            }
+          })
+        }
+        break
+      default:
+        injectedReject('Missing Authorization')
+    }
+  })
+}
+
+/***
+ * !This is an internal function that should not be called by the end user!
+
+ * Builds a URL for HTTP request
+ * @param {object} anonymous: anonyous object with arguments:
+ * @param {string} AMaaSClass: class being requested/sent (e.g. Transaction) (required)
+ * @param {string} AMId: Asset Manager Id (required)
+ * @param {string} resourceId: Id of the resource being requested (e.g. book_id)
+*/
+export function buildURL({ AMaaSClass, AMId, resourceId }) {
+  let baseURL = ''
+  switch (AMaaSClass) {
+    case 'book':
+      baseURL = `${getEndpoint()}/book/books`
+      break
+    case 'parties':
+      baseURL = `${getEndpoint()}/party/parties`
+      break
+    case 'assetManagers':
+      baseURL = `${getEndpoint()}/assetmanager/asset-managers`
+      break
+    case 'assets':
+      baseURL = `${getEndpoint()}/asset/assets`
+      break
+    case 'positions':
+      baseURL = `${getEndpoint()}/transaction/positions`
+      break
+    case 'allocations':
+      baseURL = `${getEndpoint()}/transaction/allocations`
+      break
+    case 'monitorItems':
+      baseURL = `${getEndpoint()}/monitor/items`
+      break
+    case 'monitorEvents':
+      baseURL = `${getEndpoint()}/monitor/events`
+      break
+    case 'netting':
+      baseURL = `${getEndpoint()}/transaction/netting`
+      break
+    case 'relationships':
+      baseURL = `${getEndpoint()}/assetmanager/asset-manager-relationships`
+      break
+    case 'transactions':
+      baseURL = `${getEndpoint()}/transaction/transactions`
+      break
+    case 'corporateActions':
+      baseURL = `${getEndpoint()}/corporateaction/corporate-actions`
+      break
+    default:
+      throw new Error(`Invalid class type: ${AMaaSClass}`)
+  }
+  if (!AMId) {
+    return `${baseURL}`
+  } else if (!resourceId) {
+    return `${baseURL}/${AMId}`
+  } else {
+    return `${baseURL}/${AMId}/${resourceId}`
+  }
+}
+
+export function setAuthorization() {
+  switch (stage) {
+    case 'staging':
+    case 'prod':
+    default:
+      return 'Authorization'
+  }
+}
+
+export function makeRequest({ method, url, data, query }) {
+  return getToken()
+    .then(res => {
+      switch (method) {
+        case 'GET':
+          return request.get(url).set(setAuthorization(), res).query(query)
+        case 'SEARCH':
+          return request.get(url).set(setAuthorization(), res).query(data)
+        case 'POST':
+          return request.post(url).send(data).set(setAuthorization(), res).query(query)
+        case 'PUT':
+          return request.put(url).send(data).set(setAuthorization(), res).query({ camelcase: true })
+        case 'PATCH':
+          return request.patch(url).send(data).set(setAuthorization(), res).query({ camelcase: true })
+        case 'DELETE':
+          return request.delete(url).set(setAuthorization(), res).query({ camelcase: true })
+        default:
+      }
+    })
+    .catch(err => Promise.reject(err))
+}
